# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: WIAL Backend API
  version: 1.0.0
  description: |
    OpenAPI documentation for the WIAL backend.

    Public auth is login-only. User creation is restricted to authenticated
    administrators through the managed users endpoint.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/coaches:
    get:
      tags: [Coaches]
      summary: List coaches
      operationId: listCoaches
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: chapter_id
          schema:
            type: string
        - in: query
          name: certification_level
          schema:
            type: string
        - in: query
          name: language
          schema:
            type: string
        - in: query
          name: specialization
          schema:
            type: string
      responses:
        '200':
          description: A paginated list of coaches
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoachListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Coaches]
      summary: Create a coach profile
      description: Super-admin global access; chapter leads are limited to their own chapter.
      operationId: createCoach
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachCreateRequest'
      responses:
        '201':
          description: Coach created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coach'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/coaches/{id}:
    get:
      tags: [Coaches]
      summary: Get a coach by ID
      description: Public coach profile endpoint.
      operationId: getCoach
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Coach found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coach'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Coaches]
      summary: Patch a coach profile
      operationId: patchCoach
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachPatchRequest'
      responses:
        '200':
          description: Coach patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coach'
        '400':
          description: Invalid request or no fields to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Coaches]
      summary: Delete a coach profile
      operationId: deleteCoach
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Coach deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/events:
    get:
      tags: [Events]
      summary: List events
      operationId: listEvents
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: chapter_id
          schema:
            type: string
        - in: query
          name: event_type
          schema:
            type: string
      responses:
        '200':
          description: A paginated list of events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Events]
      summary: Create an event
      description: Super-admin global access; chapter leads are limited to their own chapter.
      operationId: createEvent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreateRequest'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/events/{id}:
    patch:
      tags: [Events]
      summary: Patch an event
      operationId: patchEvent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventPatchRequest'
      responses:
        '200':
          description: Event patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid request or no fields to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Events]
      summary: Delete an event
      operationId: deleteEvent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Event deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: page_size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed or missing bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user lacks the required role or assignment scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Coach:
      type: object
      properties:
        id: { type: string }
        first_name: { type: string }
        last_name: { type: string }
        email:
          type: string
          format: email
        phone: { type: string, nullable: true }
        profile_image_url: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        specializations:
          type: array
          items: { type: string }
        languages:
          type: array
          items: { type: string }
        country: { type: string }
        city: { type: string, nullable: true }
        chapter_id: { type: string, nullable: true }
        chapter_name: { type: string, nullable: true }
        certification_level: { type: string }
        certification_date: { type: string, format: date-time }
        is_active: { type: boolean }
        linkedin_url: { type: string, nullable: true }
        website_url: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, first_name, last_name, email, specializations, languages, country, certification_level, certification_date, is_active, created_at, updated_at]
    CoachCreateRequest:
      type: object
      properties:
        first_name: { type: string }
        last_name: { type: string }
        email:
          type: string
          format: email
        phone: { type: string, nullable: true }
        profile_image_url: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        specializations:
          type: array
          items: { type: string }
        languages:
          type: array
          items: { type: string }
        country: { type: string }
        city: { type: string, nullable: true }
        chapter_id: { type: string, nullable: true }
        certification_level: { type: string }
        certification_date: { type: string, format: date-time }
        is_active: { type: boolean, default: false }
        linkedin_url: { type: string, nullable: true }
        website_url: { type: string, nullable: true }
      required: [first_name, last_name, email, country, certification_level, certification_date]
    CoachListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Coach'
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
      required: [data, page, page_size, total]
    CoachPatchRequest:
      type: object
      properties:
        first_name: { type: string }
        last_name: { type: string }
        email:
          type: string
          format: email
        phone: { type: string }
        profile_image_url: { type: string }
        bio: { type: string }
        specializations:
          type: array
          items: { type: string }
        languages:
          type: array
          items: { type: string }
        country: { type: string }
        city: { type: string }
        chapter_id: { type: string }
        certification_level: { type: string }
        certification_date: { type: string, format: date-time }
        is_active: { type: boolean }
        linkedin_url: { type: string }
        website_url: { type: string }
    Event:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        title_local: { type: string, nullable: true }
        description: { type: string, nullable: true }
        description_local: { type: string, nullable: true }
        event_type: { type: string }
        start_date: { type: string, format: date-time }
        end_date: { type: string, format: date-time, nullable: true }
        timezone: { type: string }
        location_type: { type: string }
        venue_name: { type: string, nullable: true }
        venue_address: { type: string, nullable: true }
        online_meeting_url: { type: string, nullable: true }
        chapter_id: { type: string }
        chapter_name: { type: string, nullable: true }
        max_attendees: { type: integer, nullable: true }
        current_attendees: { type: integer }
        price_amount: { type: number, format: double, nullable: true }
        price_currency: { type: string, nullable: true }
        is_free: { type: boolean }
        registration_deadline: { type: string, format: date-time, nullable: true }
        status: { type: string }
        image_url: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, title, event_type, start_date, timezone, location_type, chapter_id, current_attendees, is_free, status, created_at, updated_at]
    EventCreateRequest:
      type: object
      properties:
        title: { type: string }
        title_local: { type: string, nullable: true }
        description: { type: string, nullable: true }
        description_local: { type: string, nullable: true }
        event_type: { type: string }
        start_date: { type: string, format: date-time }
        end_date: { type: string, format: date-time, nullable: true }
        timezone: { type: string }
        location_type: { type: string }
        venue_name: { type: string, nullable: true }
        venue_address: { type: string, nullable: true }
        online_meeting_url: { type: string, nullable: true }
        chapter_id: { type: string }
        max_attendees: { type: integer, nullable: true }
        price_amount: { type: number, format: double, nullable: true }
        price_currency: { type: string, nullable: true }
        is_free: { type: boolean, default: false }
        registration_deadline: { type: string, format: date-time, nullable: true }
        status: { type: string }
        image_url: { type: string, nullable: true }
      required: [title, event_type, start_date, timezone, location_type, chapter_id]
    EventListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
      required: [data, page, page_size, total]
    EventPatchRequest:
      type: object
      properties:
        title: { type: string }
        title_local: { type: string }
        description: { type: string }
        description_local: { type: string }
        event_type: { type: string }
        start_date: { type: string, format: date-time }
        end_date: { type: string, format: date-time }
        timezone: { type: string }
        location_type: { type: string }
        venue_name: { type: string }
        venue_address: { type: string }
        online_meeting_url: { type: string }
        chapter_id: { type: string }
        max_attendees: { type: integer }
        price_amount: { type: number, format: double }
        price_currency: { type: string }
        is_free: { type: boolean }
        registration_deadline: { type: string, format: date-time }
        status: { type: string }
        image_url: { type: string }
...